arrange file structure. add code validation to updateProduct

+++ src/managers/ProductManager.js
@@ -32,9 +32,9 @@ export default class ProductManager {
   }
 
   async getProductById(req, res) {
+    res.setHeader("Content-Type", "application/json");
     let products = await this.getProducts();
     let product = products.find((product) => product.id === req.params.pid);
-    res.setHeader("Content-Type", "application/json");
     if (product) {
       res.status(200).json({ product });
     } else {
@@ -43,18 +43,21 @@ export default class ProductManager {
   }
 
   async addProduct(req, res) {
+    res.setHeader("Content-Type", "application/json");
     let { title, description, code, price, status, stock, category, thumbnails } = req.body;
     let products = await this.getProducts();
     let productExists = products.findIndex((product) => product.code === code) !== -1;
     let aFieldIsEmpty = !(title && description && code && price && stock && category);
-    res.setHeader("Content-Type", "application/json");
     if (productExists || aFieldIsEmpty) {
-      res.status(400).json({ error: `Product not added. Errors:${productExists ? " Product already exists." : ""}${aFieldIsEmpty ? " Must complete all required fields." : ""}` });
+      res.status(400).json({
+        error: `Product not added. Errors:${productExists ? " Product already exists." : ""}${
+          aFieldIsEmpty ? " Must complete all required fields." : ""
+        }`,
+      });
     } else {
       price = Number(price);
       stock = Number(stock);
-      status === "false" && (status = false);
-      status === "true" && (status = true);
+      status === "false" ? (status = false) : (status = true);
       let id = createID();
       let newProduct = new Product(id, title, description, code, price, status, stock, category, thumbnails);
       products.push(newProduct);
@@ -64,42 +67,43 @@ export default class ProductManager {
   }
 
   async updateProduct(req, res) {
+    res.setHeader("Content-Type", "application/json");
     let id = req.params.pid;
     let { title, description, code, price, status, stock, category, thumbnails } = req.body;
     let products = await this.getProducts();
-    let index = products.findIndex((product) => product.id === id);
-    let productExists = index !== -1;
-    res.setHeader("Content-Type", "application/json");
+    let indexByID = products.findIndex((product) => product.id === id);
+    let productExists = indexByID !== -1;
     if (productExists) {
-      price = Number(price);
-      stock = Number(stock);
-      if (status === "false") {
-        products[index].status = false;
+      let indexByCode = products.findIndex((product) => product.code === code);
+      let codeExists = indexByCode !== indexByID && indexByCode !== -1;
+      if (codeExists) {
+        res.status(400).json({ error: "Invalid code, already exists" });
+      } else {
+        price = Number(price);
+        stock = Number(stock);
+        status === "false" && (products[indexByID].status = false);
+        status === "true" && (products[indexByID].status = true);
+        title && (products[indexByID].title = title);
+        description && (products[indexByID].description = description);
+        code && (products[indexByID].code = code);
+        price && (products[indexByID].price = price);
+        stock && (products[indexByID].stock = stock);
+        category && (products[indexByID].category = category);
+        thumbnails && (products[indexByID].thumbnails = thumbnails);
+        await fs.promises.writeFile(this.path, JSON.stringify(products, null, 2));
+        res.status(201).json({ message: "Product updated successfully" });
       }
-      if (status === "true") {
-        products[index].status = true;
-      }
-      title && (products[index].title = title);
-      description && (products[index].description = description);
-      code && (products[index].code = code);
-      price && (products[index].price = price);
-      stock && (products[index].stock = stock);
-      category && (products[index].category = category);
-      thumbnails && (products[index].thumbnails = thumbnails);
-      await fs.promises.writeFile(this.path, JSON.stringify(products, null, 2));
-      res.status(201).json({ message: `Product updated successfully` });
     } else {
-      res.status(400).json({ error: "Product not found." });
+      res.status(400).json({ error: "Product not found" });
     }
-
   }
 
   async deleteProduct(req, res) {
+    res.setHeader("Content-Type", "application/json");
     let id = await req.params.pid;
     let products = await this.getProducts();
     let productIndex = products.findIndex((product) => product.id === id);
     let productExists = productIndex !== -1;
-    res.setHeader("Content-Type", "application/json");
     if (productExists) {
       products.splice(productIndex, 1);
       await fs.promises.writeFile(this.path, JSON.stringify(products, null, 2));
